+++ client/src/components/Settings/GoogleWalletButton.vue
@@ -18,6 +18,7 @@ import { defineProps, ref } from 'vue'
 import LoadingOverlay from '@/components/LoadingOverlay.vue'
 import i18n, { locale } from '@/helper/i18n'
 import { pulseError } from '@/script'
+import { captureError } from '@/sentry'
 
 const props = defineProps({
   disabled: {
@@ -42,7 +43,7 @@ async function onClick () {
     const data = await response.json()
 
     if (!response.ok) {
-      throw new Error(data.error || i18n('settings.passport.wallet.generate_error'))
+      throw captureError(data.error || i18n('settings.passport.wallet.generate_error'))
     }
 
     window.location.href = data.url